requirements.txt
# Core PyTorch dependencies
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# Data loading and processing
numpy>=1.21.0
Pillow>=8.3.0
opencv-python>=4.5.0

# Progress bars and utilities
tqdm>=4.62.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Image processing and metrics
scikit-image>=0.18.0
scipy>=1.7.0

# Additional utilities
pathlib2>=2.3.0
argparse

# Logging
logging

# Math and random
math
random
functools

# Typing support
typing-extensions>=4.0.0

# Optional: For better performance
# Install these if you have CUDA support
# torch-scatter
# torch-sparse
# torch-cluster
# torch-spline-conv

# Optional: For advanced image processing
# albumentations>=1.1.0
# imgaug>=0.4.0

# Optional: For logging and monitoring
# tensorboard>=2.7.0
# wandb>=0.12.0

# Optional: For development and testing
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.8.0
# mypy>=0.910 